requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.0
attrs==25.4.0
bcrypt==5.0.0
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.2
curl_cffi==0.13.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.119.0
frozendict==2.4.6
frozenlist==1.8.0
h11==0.16.0
idna==3.11
motor==3.7.1
multidict==6.7.0
multitasking==0.0.12
numpy==2.3.3
pandas==2.3.3
peewee==3.18.2
platformdirs==4.5.0
propcache==0.4.1
protobuf==6.32.1
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.0
pydantic-settings==2.11.0
pydantic_core==2.41.1
pymongo==4.15.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
pytz==2025.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.37.0
websockets==15.0.1
yarl==1.22.0
yfinance==0.2.66
